@@ -4,11 +4,12 @@
   "description": "Highlight select fragments of texts",
   "main": "lib/highlighter.js",
   "peerDependencies": {
-    "react": ">= 0.12.0"
+    "react": "^0.14.0"
   },
   "scripts": {
-    "test": "./node_modules/.bin/mocha",
+    "test": "./node_modules/.bin/_mocha",
     "coverage": "./node_modules/.bin/istanbul cover ./node_modules/.bin/_mocha -- -R spec",
+    "check-coverage": "./node_modules/.bin/istanbul check-coverage --statement 90 --function 100",
     "dist": "./node_modules/.bin/webpack"
   },
   "jest": {
@@ -33,9 +34,13 @@
   "homepage": "https://github.com/helior/react-highlighter",
   "devDependencies": {
     "chai": "^2.1.2",
-    "istanbul": "^0.3.13",
+    "istanbul": "^0.4.1",
     "jsdom": "^3.1.2",
-    "mocha": "^2.2.1",
+    "mocha": "^2.3.4",
+    "react-dom": "~0.14.3",
     "webpack": "^1.7.3"
+  },
+  "dependencies": {
+    "escape-string-regexp": "^1.0.4"
   }
 }
