@@ -1,5 +1,6 @@
 var React = require('react');
 var RegExpPropType = require('./regExpPropType');
+var escapeStringRegexp = require('escape-string-regexp');
 
 
 var Highlighter = React.createClass({displayName: "Highlighter",
@@ -14,14 +15,16 @@ var Highlighter = React.createClass({displayName: "Highlighter",
     ]).isRequired,
     caseSensitive: React.PropTypes.bool,
     matchElement: React.PropTypes.string,
-    matchClass: React.PropTypes.string
+    matchClass: React.PropTypes.string,
+    matchStyle: React.PropTypes.object
   },
 
   getDefaultProps: function() {
     return {
       caseSensitive: false,
       matchElement: 'strong',
-      matchClass: 'highlight'
+      matchClass: 'highlight',
+      matchStyle: {}
     }
   },
 
@@ -82,7 +85,12 @@ var Highlighter = React.createClass({displayName: "Highlighter",
       flags +='i';
     }
 
-    return new RegExp(this.props.search, flags);
+    var search = this.props.search;
+    if (typeof this.props.search === 'string') {
+      search = escapeStringRegexp(search);
+    }
+
+    return new RegExp(search, flags);
   },
 
   /**
@@ -171,7 +179,11 @@ var Highlighter = React.createClass({displayName: "Highlighter",
    */
   renderHighlight: function(string) {
     this.count++;
-    return React.DOM[this.props.matchElement]({'key': this.count, 'className': this.props.matchClass}, string);
+    return React.DOM[this.props.matchElement]({
+      key: this.count,
+      className: this.props.matchClass,
+      style: this.props.matchStyle
+    }, string);
   }
 });
 
